+++ client/kmp/composeApp/src/commonMain/kotlin/org/calypsonet/keyple/demo/reload/remote/Datastore.kt
@@ -1,0 +1,27 @@
+/* **************************************************************************************
+ * Copyright (c) 2024 Calypso Networks Association https://calypsonet.org/
+ *
+ * See the NOTICE file(s) distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ************************************************************************************** */
+package org.calypsonet.keyple.demo.reload.remote
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import okio.Path.Companion.toPath
+
+internal const val dataStoreFileName = "keyple_client.preferences_pb"
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+expect class DataStorePathProducer {
+  fun producePath(): String
+}
+
+fun createDataStore(pathProducer: DataStorePathProducer): DataStore<Preferences> =
+    PreferenceDataStoreFactory.createWithPath(produceFile = { pathProducer.producePath().toPath() })