+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added the `GET` REST resource `/export-card-selection-scenario` to allow the client to retrieve an exported card
+  selection scenario.
+- Added server-side management of the `initialCardContent` JSON property, to enable import of the processed card 
+  selection scenario. 
 
 ## [2025.05.23]
 ### Fixed